第二次上传项目的Mapper类（郭鹏生）

+++ src/main/java/cn/kgc/ybis/pojo/Download.java
@@ -1,17 +1,17 @@
 package cn.kgc.ybis.pojo;
 
 public class Download {
-    private Long downId;
+    private Integer downId;
 
     private String tTitle;
 
     private String downPath;
 
-    public Long getDownId() {
+    public Integer getDownId() {
         return downId;
     }
 
-    public void setDownId(Long downId) {
+    public void setDownId(Integer downId) {
         this.downId = downId;
     }
